job-board-api-service/requirements.txt
FastAPI
Uvicorn[standard]
SQLAlchemy
python-dotenv
psycopg2-binary
pydantic
pytest
testcontainers[postgresql]
fastapi_keycloak_middleware
starlette
itsdangerous
PyJWT
requests
python-jose[cryptography]
httpx
playwright
asyncpg